+++ kinder-models/tests/dynamic3d/ground/test_ground_parameterized_skills.py
@@ -15,10 +15,10 @@ from relational_structs.utils import create_state_from_dict
 from spatialmath import SE2
 
 from kinder_models.dynamic3d.ground.parameterized_skills import (
-    PyBulletSim,
     create_lifted_controllers,
     get_target_robot_pose_from_parameters,
 )
+from kinder_models.dynamic3d.utils import PyBulletSim
 
 kinder.register_all_environments()
 
@@ -1404,200 +1404,5 @@ def test_pick_ground_toss():
     print("cube_orientation", cube_orientation)
     print("robot base position", robot_base_position)
     print("distance", distance)
-
-    env.close()
-
-
-def test_open_drawer():
-    """Test open drawer."""
-
-    # Create the environment.
-    num_cubes = 5
-    env = kinder.make(
-        f"kinder/SweepIntoDrawer3D-o{num_cubes}-v0", render_mode="rgb_array"
-    )
-    if MAKE_VIDEOS:
-        env = RecordVideo(
-            env, "unit_test_videos", name_prefix=f"TidyBot3D-cupboard-o{num_cubes}-real"
-        )
-
-    # Reset the environment and get the initial state.
-    obs, _ = env.reset(seed=123)
-    for _ in range(5):
-        obs, _, _, _, _ = env.step(np.zeros(11))
-    assert isinstance(env.observation_space, ObjectCentricBoxSpace)
-    state = env.observation_space.devectorize(obs)
-
-    assert state is not None
-    pybullet_sim = PyBulletSim(state, rendering=False)
-
-    controllers = create_lifted_controllers(env.action_space, pybullet_sim=pybullet_sim)
-
-    # create the pick ground controller.
-    lifted_controller = controllers["open_drawer"]
-    robot = _get_robot_from_state(state)
-    drawer = state.get_object_from_name("wiper_0")
-    object_parameters = (robot, drawer)
-    controller = lifted_controller.ground(object_parameters)
-    # params = controller.sample_parameters(state, np.random.default_rng(123))
-    params = np.array([0.7, -np.pi])
-
-    # Reset and execute the controller until it terminates.
-    controller.reset(state, params)
-    for _ in range(300):
-        action = controller.step()
-        obs, _, _, _, _ = env.step(action)
-        next_state = env.observation_space.devectorize(obs)
-        controller.observe(next_state)
-        state = next_state
-        if controller.terminated():
-            break
-    else:
-        assert False, "Controller did not terminate"
-
-    env.close()
-
-
-def test_pick_wiper():
-    """Test pick wiper."""
-
-    # Create the environment.
-    num_cubes = 5
-    env = kinder.make(
-        f"kinder/SweepIntoDrawer3D-o{num_cubes}-v0", render_mode="rgb_array"
-    )
-    if MAKE_VIDEOS:
-        env = RecordVideo(
-            env, "unit_test_videos", name_prefix=f"TidyBot3D-cupboard-o{num_cubes}-real"
-        )
-
-    # Reset the environment and get the initial state.
-    obs, _ = env.reset(seed=123)
-    for _ in range(5):
-        obs, _, _, _, _ = env.step(np.zeros(11))
-    assert isinstance(env.observation_space, ObjectCentricBoxSpace)
-    state = env.observation_space.devectorize(obs)
-
-    assert state is not None
-    pybullet_sim = PyBulletSim(state, rendering=False)
-
-    controllers = create_lifted_controllers(env.action_space, pybullet_sim=pybullet_sim)
-
-    # create the pick ground controller.
-    lifted_controller = controllers["pick_wiper"]
-    robot = _get_robot_from_state(state)
-    drawer = state.get_object_from_name("wiper_0")
-    object_parameters = (robot, drawer)
-    controller = lifted_controller.ground(object_parameters)
-    # params = controller.sample_parameters(state, np.random.default_rng(123))
-    params = np.array([0.7, -np.pi])
-
-    # Reset and execute the controller until it terminates.
-    controller.reset(state, params)
-    for _ in range(300):
-        action = controller.step()
-        obs, _, _, _, _ = env.step(action)
-        next_state = env.observation_space.devectorize(obs)
-        controller.observe(next_state)
-        state = next_state
-        if controller.terminated():
-            break
-    else:
-        assert False, "Controller did not terminate"
-
-    env.close()
-
-
-def test_open_drawer_pick_sweep_wiper():
-    """Test open drawer, pick and sweep wiper."""
-
-    # Create the environment.
-    num_cubes = 5
-    env = kinder.make(
-        f"kinder/SweepIntoDrawer3D-o{num_cubes}-v0", render_mode="rgb_array"
-    )
-    if MAKE_VIDEOS:
-        env = RecordVideo(
-            env, "unit_test_videos", name_prefix=f"TidyBot3D-cupboard-o{num_cubes}-real"
-        )
-
-    # Reset the environment and get the initial state.
-    obs, _ = env.reset(seed=123)
-    for _ in range(5):
-        obs, _, _, _, _ = env.step(np.zeros(11))
-    assert isinstance(env.observation_space, ObjectCentricBoxSpace)
-    state = env.observation_space.devectorize(obs)
-
-    assert state is not None
-    pybullet_sim = PyBulletSim(state, rendering=False)
-
-    controllers = create_lifted_controllers(env.action_space, pybullet_sim=pybullet_sim)
-
-    # create the pick ground controller.
-    lifted_controller = controllers["open_drawer"]
-    robot = _get_robot_from_state(state)
-    drawer = state.get_object_from_name("wiper_0")
-    object_parameters = (robot, drawer)
-    controller = lifted_controller.ground(object_parameters)
-    # params = controller.sample_parameters(state, np.random.default_rng(123))
-    params = np.array([0.7, -np.pi])
-
-    # Reset and execute the controller until it terminates.
-    controller.reset(state, params)
-    for _ in range(300):
-        action = controller.step()
-        obs, _, _, _, _ = env.step(action)
-        next_state = env.observation_space.devectorize(obs)
-        controller.observe(next_state)
-        state = next_state
-        if controller.terminated():
-            break
-    else:
-        assert False, "Controller did not terminate"
-
-    # create the pick ground controller.
-    lifted_controller = controllers["pick_wiper"]
-    robot = _get_robot_from_state(state)
-    drawer = state.get_object_from_name("wiper_0")
-    object_parameters = (robot, drawer)
-    controller = lifted_controller.ground(object_parameters)
-    # params = controller.sample_parameters(state, np.random.default_rng(123))
-    params = np.array([0.7, -np.pi])
-
-    # Reset and execute the controller until it terminates.
-    controller.reset(state, params)
-    for _ in range(300):
-        action = controller.step()
-        obs, _, _, _, _ = env.step(action)
-        next_state = env.observation_space.devectorize(obs)
-        controller.observe(next_state)
-        state = next_state
-        if controller.terminated():
-            break
-    else:
-        assert False, "Controller did not terminate"
-
-    # create the place ground controller.
-    lifted_controller = controllers["sweep"]
-    robot = _get_robot_from_state(state)
-    wiper = state.get_object_from_name("wiper_0")
-    target_cube = state.get_object_from_name("cube_0")
-    object_parameters = (robot, wiper, target_cube)
-    controller = lifted_controller.ground(object_parameters)
-    # params = controller.sample_parameters(state, np.random.default_rng(123))
-    params = np.array([0.55, -np.pi])
-
-    # Reset and execute the controller until it terminates.
-    controller.reset(state, params)
-    for _ in range(200):
-        action = controller.step()
-        obs, _, _, _, _ = env.step(action)
-        next_state = env.observation_space.devectorize(obs)
-        controller.observe(next_state)
-        state = next_state
-        if controller.terminated():
-            break
-    else:
-        assert False, "Controller did not terminate"
 
     env.close()
